% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_table_row.R
\name{delete_table_row}
\alias{delete_table_row}
\title{Delete a row from a table}
\usage{
delete_table_row(
  table_name = NULL,
  id_value = NULL,
  id_column = "id",
  schema = Sys.getenv("SUPABASE_SCHEMA"),
  conn = make_connection()
)
}
\arguments{
\item{table_name}{The name of the table.}

\item{id_value}{The ID value of the row to delete.}

\item{id_column}{The ID column name.}

\item{schema}{The schema name.}

\item{conn}{A database connection object.}
}
\description{
Delete a row from a table
}
