% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_api_available}
\alias{.check_api_available}
\title{Check that API credentials are available}
\usage{
.check_api_available()
}
\description{
Check that API credentials are available
}
\keyword{internal}
