% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.sb_api_abort}
\alias{.sb_api_abort}
\title{Abort with a structured PostgREST error message}
\usage{
.sb_api_abort(response)
}
\arguments{
\item{response}{An httr2 response object with a non-2xx status}
}
\description{
Abort with a structured PostgREST error message
}
\keyword{internal}
