% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_table.R, R/supabase.R
\name{empty_table}
\alias{empty_table}
\title{Delete all rows from table}
\usage{
empty_table(table_name = NULL, schema = "public", conn = make_connection())

empty_table(table_name = NULL, schema = "public", conn = make_connection())
}
\arguments{
\item{table_name}{The name of the table.}

\item{schema}{The schema name.}

\item{conn}{A database connection object.}
}
\description{
Delete all rows from table

Delete all rows from table
}
