% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_columns.R
\name{filter_columns}
\alias{filter_columns}
\title{Filter columns based on schema and operation type}
\usage{
filter_columns(schema_info, is_update = FALSE)
}
\arguments{
\item{schema_info}{A data frame containing schema information.}

\item{is_update}{Logical indicating whether the operation is an update.}
}
\value{
A vector of column names to include.
}
\description{
Filter columns based on schema and operation type
}
