% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_db_connect.R
\name{get_connection}
\alias{get_connection}
\title{Get the current connection}
\usage{
get_connection()
}
\value{
The current database connection
}
\description{
Get the current connection
}
\keyword{internal}
