% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cron_time.R
\name{get_cron_time}
\alias{get_cron_time}
\title{Generate the CRON time}
\usage{
get_cron_time(time = Sys.time(), tz = "Asia/Kolkata")
}
\arguments{
\item{time}{The time for which the CRON time is to be generated}

\item{tz}{The timezone of the time}
}
\value{
The CRON time
}
\description{
Generate the CRON time
}
