% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_freeform_table_query.R
\name{get_freeform_table_query}
\alias{get_freeform_table_query}
\title{Query data from a table with a freeform text query}
\usage{
get_freeform_table_query(
  query = NULL,
  schema = Sys.getenv("SUPABASE_SCHEMA"),
  conn = make_connection()
)
}
\arguments{
\item{query}{The custom query as text}

\item{schema}{The schema name.}

\item{conn}{A database connection object.}
}
\value{
A data frame with table data.
}
\description{
Query data from a table with a freeform text query
}
