% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_key.R
\name{get_latest_key}
\alias{get_latest_key}
\title{Get the latest key value for a table}
\usage{
get_latest_key(
  table_name = NULL,
  id_column = "id",
  schema = "public",
  conn = make_connection()
)
}
\arguments{
\item{table_name}{The name of the table.}

\item{id_column}{The ID column name.}

\item{schema}{The schema name.}

\item{conn}{A database connection object.}
}
\value{
The latest key value plus one.
}
\description{
Get the latest key value for a table
}
