% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_data.R
\name{get_table_data}
\alias{get_table_data}
\title{Retrieve data from a table}
\usage{
get_table_data(
  table_name = NULL,
  limit = 0,
  schema = Sys.getenv("SUPABASE_SCHEMA"),
  conn = make_connection()
)
}
\arguments{
\item{table_name}{The name of the table.}

\item{limit}{The number of rows to retrieve.}

\item{schema}{The schema name.}

\item{conn}{A database connection object.}
}
\value{
A data frame with table data.
}
\description{
Retrieve data from a table
}
