% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_query.R
\name{get_table_query}
\alias{get_table_query}
\title{Query data from a table}
\usage{
get_table_query(
  table_name = NULL,
  columns = list("*"),
  filter_query = list(),
  schema = Sys.getenv("SUPABASE_SCHEMA"),
  conn = make_connection()
)
}
\arguments{
\item{table_name}{The name of the table.}

\item{columns}{The columns spread as a list. Default is list("*")}

\item{filter_query}{The custom query spread as list}

\item{schema}{The schema name.}

\item{conn}{A database connection object.}
}
\value{
A data frame with table data.
}
\description{
Query data from a table
}
