% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_schema.R
\name{get_table_schema}
\alias{get_table_schema}
\title{Retrieve the schema of a table}
\usage{
get_table_schema(
  table_name = NULL,
  schema = "public",
  conn = make_connection()
)
}
\arguments{
\item{table_name}{The name of the table.}

\item{schema}{The schema name.}

\item{conn}{A database connection object.}
}
\value{
A data frame with table schema details.
}
\description{
Retrieve the schema of a table
}
