% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_type_mapping.R
\name{get_type_mapping}
\alias{get_type_mapping}
\title{Get the mapping of R types to PostgreSQL types}
\usage{
get_type_mapping()
}
\value{
A named list mapping R types to PostgreSQL types
}
\description{
Get the mapping of R types to PostgreSQL types
}
