% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_table.R
\name{is_valid_table}
\alias{is_valid_table}
\title{Check if the table exists}
\usage{
is_valid_table(table_name = NULL, conn = make_connection())
}
\arguments{
\item{table_name}{The name of the table.}

\item{conn}{A database connection object.}
}
\value{
TRUE if the table exists, FALSE otherwise.
}
\description{
Check if the table exists
}
