% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_connection.R
\name{make_connection}
\alias{make_connection}
\title{Create a database connection}
\usage{
make_connection(supabase_creds = read_supabase_creds())
}
\arguments{
\item{supabase_creds}{A list of Supabase credentials.}
}
\value{
A database connection object.
}
\description{
Create a database connection
}
