% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_sql_to_r.R
\name{map_sql_to_r}
\alias{map_sql_to_r}
\title{Map SQL data types to R data types}
\usage{
map_sql_to_r(data_type, type_mapping = get_type_mapping())
}
\arguments{
\item{data_type}{SQL data type as a character string.}

\item{type_mapping}{A named list mapping SQL data types to R data types.}
}
\value{
Corresponding R data type as a character string.
}
\description{
Map SQL data types to R data types
}
