% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_table_row.R
\name{put_table_row}
\alias{put_table_row}
\title{Insert or update a table row}
\usage{
put_table_row(
  table_name = NULL,
  input_list = list(),
  is_update = FALSE,
  schema = "hrafnagud",
  conn = make_connection()
)
}
\arguments{
\item{table_name}{The name of the table.}

\item{input_list}{A list of column-value pairs.}

\item{is_update}{Whether the operation is an update.}

\item{schema}{The schema name.}

\item{conn}{A database connection object.}
}
\description{
Insert or update a table row
}
