% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_supabase_creds.R
\name{read_supabase_creds}
\alias{read_supabase_creds}
\title{Read Supabase credentials from environment variables}
\usage{
read_supabase_creds()
}
\value{
A list of Supabase credentials.
}
\description{
Read Supabase credentials from environment variables
}
