% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_api_read.R
\name{sb_api_read}
\alias{sb_api_read}
\title{Read table data via API}
\usage{
sb_api_read(table = NULL, where = NULL, limit = NULL)
}
\arguments{
\item{table}{The table name}

\item{where}{A list of filter conditions (e.g. list(id = 1
or list(created_at = list(gt = "2024-01-01")))}

\item{limit}{Maximum rows to return (NULL for all)}
}
\value{
A \code{tibble}
}
\description{
Read table data via API
}
