% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_api_status.R
\name{sb_api_status}
\alias{sb_api_status}
\title{Check API backend status}
\usage{
sb_api_status()
}
\value{
A list with API connection info
}
\description{
Check API backend status
}
\examples{
\dontrun{
sb_api_connect()

sb_api_status()
}
}
