% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_db_creds.R
\name{sb_db_creds}
\alias{sb_db_creds}
\title{Read Supabase credentials from environment variables}
\usage{
sb_db_creds()
}
\value{
A list of Supabase credentials
}
\description{
Read Supabase credentials from environment variables
}
