% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_db_insert.R
\name{sb_db_insert}
\alias{sb_db_insert}
\title{Insert rows into a table}
\usage{
sb_db_insert(table = NULL, data = NULL, schema = get_schema())
}
\arguments{
\item{table}{The table name}

\item{data}{A data frame of rows to insert}

\item{schema}{The schema name}
}
\value{
Number of rows inserted (invisibly)
}
\description{
Insert rows into a table
}
