% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_db_schema.R
\name{sb_db_schema}
\alias{sb_db_schema}
\title{Get table schema}
\usage{
sb_db_schema(table = NULL, schema = get_schema())
}
\arguments{
\item{table}{The table name}

\item{schema}{The schema name}
}
\value{
A data frame with column_name and data_type
}
\description{
Get table schema
}
