% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_db_truncate.R
\name{sb_db_truncate}
\alias{sb_db_truncate}
\title{Truncate a table}
\usage{
sb_db_truncate(table = NULL, schema = get_schema())
}
\arguments{
\item{table}{The table name}

\item{schema}{The schema name}
}
\value{
Invisible NULL
}
\description{
Truncate a table
}
