% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_wrappers.R
\name{sb_delete}
\alias{sb_delete}
\title{Delete rows from a table}
\usage{
sb_delete(table = NULL, where = NULL, schema = get_schema())
}
\arguments{
\item{table}{The table name}

\item{where}{A named list for WHERE clause}

\item{schema}{The schema name}
}
\value{
Number of rows deleted (invisibly)
}
\description{
Delete rows from a table
}
