% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_wrappers.R
\name{sb_query}
\alias{sb_query}
\title{Query a table}
\usage{
sb_query(
  table = NULL,
  columns = "*",
  where = NULL,
  limit = 0,
  sql = NULL,
  schema = get_schema()
)
}
\arguments{
\item{table}{The table name}

\item{columns}{Columns to select}

\item{where}{A named list for WHERE clause}

\item{limit}{Maximum rows to return}

\item{sql}{Raw SQL query (db backend only)}

\item{schema}{The schema name}
}
\value{
A data frame with query results
}
\description{
Query a table
}
