% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_wrappers.R
\name{sb_schema}
\alias{sb_schema}
\title{Get table schema}
\usage{
sb_schema(table = NULL, schema = get_schema())
}
\arguments{
\item{table}{The table name}

\item{schema}{The schema name}
}
\value{
A data frame with column info
}
\description{
Get table schema
}
