% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_wrappers.R
\name{sb_table_exists}
\alias{sb_table_exists}
\title{Check if table exists}
\usage{
sb_table_exists(table = NULL, schema = get_schema())
}
\arguments{
\item{table}{The table name}

\item{schema}{The schema name}
}
\value{
TRUE if table exists
}
\description{
Check if table exists
}
