% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_wrappers.R
\name{sb_tables}
\alias{sb_tables}
\title{List tables}
\usage{
sb_tables(schema = get_schema())
}
\arguments{
\item{schema}{The schema name}
}
\value{
A character vector of table names
}
\description{
List tables
}
