% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_wrappers.R
\name{sb_upsert}
\alias{sb_upsert}
\title{Upsert rows into a table}
\usage{
sb_upsert(
  table = NULL,
  data = NULL,
  conflict_columns = NULL,
  schema = get_schema()
)
}
\arguments{
\item{table}{The table name}

\item{data}{A data frame of rows to upsert}

\item{conflict_columns}{Column(s) to check for conflicts}

\item{schema}{The schema name}
}
\value{
Number of rows affected (invisibly)
}
\description{
Upsert rows into a table
}
