% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_db_connect.R
\name{sb_db_connect}
\alias{sb_db_connect}
\title{Connect to Supabase}
\usage{
sb_db_connect(schema = Sys.getenv("SUPABASE_SCHEMA"))
}
\arguments{
\item{schema}{The schema name}
}
\value{
Invisible connection object
}
\description{
Connect to Supabase
}
