% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_db_read.R
\name{sb_db_read}
\alias{sb_db_read}
\title{Read table data}
\usage{
sb_db_read(table = NULL, limit = 0, schema = get_schema())
}
\arguments{
\item{table}{The table name}

\item{limit}{Maximum rows to return (0 for all)}

\item{schema}{The schema name}
}
\value{
A data frame with table data
}
\description{
Read table data
}
