% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_db_tables.R
\name{sb_db_tables}
\alias{sb_db_tables}
\title{List tables in schema}
\usage{
sb_db_tables(schema = get_schema())
}
\arguments{
\item{schema}{The schema name}
}
\value{
A character vector of table names
}
\description{
List tables in schema
}
\examples{
\dontrun{
# List all tables in current schema
sb_db_tables()

# List tables in a specific schema
sb_db_tables(schema = "auth")
}
}
